move TabView under the composers directory
